requirements.txt
streamlit
google-generativeai
gtts
SpeechRecognition
pydub
python-dotenv
deep-translator